feature: use user-provided lookup fns for deserialization
The initial design allowed the use to provide a Map to define
free variables (either as What4 Exprs or SymFns). This change
now instead uses user-provided lookup functions whose codomain
is an IO value. This means that standard users can basically just
wrap a simple map in a lookup, but users with more complicated needs
can hide that complexity in the provided function. E.g., semmc generates
s-expressions which use a keyword like `undefined` to mean a fresh,
undefined bitvector. By using a lookup function, we allow them to
generate those fresh bitvectors whenever we encounter and lookup
the meaning of the free identifier `undefined`.