Add support for ArrayMap and some float operations
This commit adds support for serializing and deserializing the `ArrayMap`
constructor from what4, which is now generated more often due to various
optimizations in what4.

It also adds support for the Float type, including serialization/deserialization
of literals and two unary operations (as a proof-of-concept).

Note that the tests currently non-determinstically fail due to some strange
behavior in libBF